Move config module to top level
